Luciana: atualizando IdConsultor no cliente

+++ consultoria/src/main/java/br/com/squad04/consultoria/repository/ClientesRepo.java
@@ -3,5 +3,5 @@ package br.com.squad04.consultoria.repository;
 import br.com.squad04.consultoria.model.Clientes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientesRepo extends JpaRepository<Clientes, Integer> {
+public interface ClientesRepo extends JpaRepository<Clientes, Long> {
 }
